Fix the parameters of a <dist_spec>

If the given <dist_spec> has any uncertainty, it is removed and the corresponding distribution converted into a fixed one.

# S3 method for class 'dist_spec'
fix_parameters(x, strategy = c("mean", "sample"), ...)

Arguments

x

A <dist_spec>

strategy

Character; either "mean" (use the mean estimates of the mean and standard deviation) or "sample" (randomly sample mean and standard deviation from uncertainty given in the <dist_spec>

...

ignored

Value

A <dist_spec> object without uncertainty

Examples

# An uncertain gamma distribution with mean 3 and sd 2
dist <- LogNormal(
  meanlog = Normal(3, 0.5), sdlog = Normal(2, 0.5), max = 20
)

fix_parameters(dist)
#> - lognormal distribution (max: 20):
#>   meanlog:
#>     3
#>   sdlog:
#>     2