Creates a delay distribution as the sum of two other delay distributions.

# S3 method for class 'dist_spec'
e1 + e2

Arguments

e1

The first delay distribution (of type <dist_spec>) to combine.

e2

The second delay distribution (of type <dist_spec>) to combine.

Value

A delay distribution representing the sum of the two delays

Examples

# A fixed lognormal distribution with mean 5 and sd 1.
dist1 <- LogNormal(
  meanlog = 1.6, sdlog = 1, max = 20
)
dist1 + dist1
#> Composite distribution:
#> - lognormal distribution (max: 20):
#>   meanlog:
#>     1.6
#>   sdlog:
#>     1
#> - lognormal distribution (max: 20):
#>   meanlog:
#>     1.6
#>   sdlog:
#>     1

# An uncertain gamma distribution with mean 3 and sd 2
dist2 <- Gamma(
  mean = Normal(3, 0.5), sd = Normal(2, 0.5), max = 20
)
#> Warning: ! Uncertain gamma distribution specified in terms of parameters that are not
#>   the "natural" parameters of the distribution shape and rate.
#> ℹ Converting using a crude and very approximate method that is likely to
#>   produce biased results.
#> ℹ If possible it is preferable to specify the distribution directly in terms of
#>   the natural parameters.
dist1 + dist2
#> Composite distribution:
#> - lognormal distribution (max: 20):
#>   meanlog:
#>     1.6
#>   sdlog:
#>     1
#> - gamma distribution (max: 20):
#>   shape:
#>     - normal distribution:
#>       mean:
#>         2.2
#>       sd:
#>         0.48
#>   rate:
#>     - normal distribution:
#>       mean:
#>         0.75
#>       sd:
#>         0.28